Restructure project and remove unused files
Reorganized project to move src files to a `lib` directory, updated Gradle build files, and removed unused Python and other source files. Adjusted `.gitignore` and Gradle properties for consistency and updated several Kotlin class paths.
