+++ .gitignore
@@ -1,8 +1,8 @@
 .gradle
 build/
 !gradle/wrapper/gradle-wrapper.jar
-!**/src/main/**/build/
-!**/src/test/**/build/
+!**/src/*main/**/build/
+!**/src/*test/**/build/
 
 ### STS ###
 .apt_generated
@@ -22,8 +22,8 @@ bin/
 *.iml
 *.ipr
 out/
-!**/src/main/**/out/
-!**/src/test/**/out/
+!**/src/*main/**/out/
+!**/src/*test/**/out/
 
 ### NetBeans ###
 /nbproject/private/
@@ -34,3 +34,5 @@ out/
 
 ### VS Code ###
 .vscode/
+# Ignore Gradle build output directory
+build
