+++ gradle.properties
@@ -1,4 +1,6 @@
-kotlin.code.style=official
-ktorVersion=2.3.12
-kotlinVersion=2.0.21
-logbackVersion=1.5.6
+# This file was generated by the Gradle 'init' task.
+# https://docs.gradle.org/current/userguide/build_environment.html#sec:gradle_configuration_properties
+
+org.gradle.parallel=true
+org.gradle.caching=true
+
