+++ lib/src/commonMain/kotlin/ai/solace/core/CORE_ARCHITECTURE.md
