+++ lib/src/commonMain/kotlin/ai/solace/core/actor/state/ActorState.kt
