+++ lib/src/commonMain/resources/openapi/documentation.yaml
