+++ lib/src/commonMain/resources/scripts/HelloWorldActor.kts
