--- src/README.md
@@ -1,16 +1,0 @@
-# src
-
-## Overview
-The `src` directory contains the source code for the SolaceCore framework, including core components, actor implementations, and examples.
-
-## Structure
-- **main**: Contains the main source code for the framework.
-- **test**: Contains test cases for the framework components.
-
-## TODO
-- Implement additional actor examples.
-- Develop test cases for new features.
-
-## Completed
-- Core actor and supervisor classes.
-- Initial set of example actors.
