--- src/jvmMain/ai/solace/core/WorkflowManager.kt
@@ -1,2 +1,0 @@
-package ai.solace.core 
-
