--- src/macosArm64NativeMain/ai/solace/core/actor/types/NativeActor.kt
@@ -1,2 +1,0 @@
-package ai.solace.core.actor.types 
-
