--- src/resources/README.md
@@ -1,15 +1,0 @@
-# resources
-
-## Overview
-The `resources` directory contains additional resources for the SolaceCore framework, such as configuration files and scripts.
-
-## Structure
-- **openapi**: OpenAPI specifications for API endpoints.
-- **scripts**: Scripts for various tasks and configurations.
-
-## TODO
-- Add configuration files for new features.
-- Develop scripts for automated tasks.
-
-## Completed
-- Initial set of configuration files and scripts.
