--- src/resources/openapi/documentation.yaml
@@ -1,27 +1,0 @@
-openapi: "3.1.0"
-info:
-  title: "SolaceCore API"
-  description: "SolaceCore API"
-  version: "1.0.0"
-servers:
-  - url: "http://localhost:8080"
-paths:
-  /customer/{id}:
-    get:
-      summary: "Get a customer by ID"
-      parameters:
-        - in: path
-          name: id
-          schema:
-            type: integer
-          required: true
-          description: "ID of the customer to retrieve"
-      responses:
-        '200':
-          description: "Successful response"
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Customer"
-        '404':
-          description: "Customer not found"
